+++ src/main.rs
@@ -1,6 +1,7 @@
 mod sio;
 
-use std::{collections::HashMap, io::Write, path::Path, result::Result, sync::Mutex, time::Duration};
+use color_eyre::eyre::Result;
+use std::{collections::HashMap, io::Write, path::Path, sync::Mutex, time::Duration};
 
 #[macro_use]
 extern crate log;
@@ -29,6 +30,8 @@ lazy_static! {
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
+  color_eyre::install()?;
+
   let app = App::new("").version(env!("CARGO_PKG_VERSION")).author(env!("CARGO_PKG_AUTHORS")).about(env!("CARGO_PKG_DESCRIPTION")).arg(Arg::with_name("interval").short('i').long("interval").env("INTERVAL").required(false).default_value("60").help("Refresh interval in seconds")).arg(Arg::with_name("cfg_path").short('c').long("cfg_path").env("CFG_PATH").required(false).default_value("cfg").help("Configuration path")).arg(Arg::with_name("ip").short('h').long("ip").env("IP").required(true)).arg(Arg::with_name("auth_usr").short('u').long("auth_usr").env("AUTH_USR").required(true)).arg(Arg::with_name("auth_pwd").short('p').long("auth_pwd").env("AUTH_PWD").requires("auth_usr").required(true)).arg(Arg::with_name("v").short('v').multiple(true).takes_value(false).required(false).help("Log verbosity (-v, -vv, -vvv...)")).get_matches();
 
   match app.occurrences_of("v") {